feat: use new advanced build urls to replace custom plugin
