+++ packages/webpack/src/webpack.ts
@@ -8,8 +8,8 @@ import type { Compiler, Watching } from 'webpack'
 import type { Nuxt } from '@nuxt/schema'
 import { joinURL } from 'ufo'
 import { logger, useNuxt } from '@nuxt/kit'
-import { DynamicBasePlugin } from '../../vite/src/plugins/dynamic-base'
 import { composableKeysPlugin } from '../../vite/src/plugins/composable-keys'
+import { DynamicBasePlugin } from './plugins/dynamic-base'
 import { createMFS } from './utils/mfs'
 import { registerVirtualModules } from './virtual-modules'
 import { client, server } from './configs'
